Refactoring: use pattern-matching switch for cases with null
This is the biggest new feature of Java 21
and allows us to switch over types as done in this commit.
No pattern matching with deconstruction of records is done yet.
Also, "case null" is allowed now.

This commit changes only cases where there was already a branch for null
in the if-then-else chain.
While we are at it, also add @Nullable for inputs and outputs in these
cases.

Commit based on refactoring by Eclipse with some manual adjustments.