Refactoring: use pattern-matching switch for cases where default throws
This is the biggest new feature of Java 21
and allows us to switch over types as done in this commit.
No pattern matching with deconstruction of records is done yet.

This commit changes only cases where the existing code throws an
exception in the final "else" branch and has no branch for "null".
In the new code we omit a case for "null", so it would trigger a
NullPointerException.
This can mean that the new code throws a different exception than the
old code for null, but this should not be a problem.
All the cases in the old code are unchecked exceptions
or it looks as if the exception thrown would not be a proper handling
of an expected null anyway.

Commit based on refactoring by Eclipse with some manual adjustments.