Refactoring: use pattern-matching switch for cases where null throws
This is the biggest new feature of Java 21
and allows us to switch over types as done in this commit.
No pattern matching with deconstruction of records is done yet.

This commit changes only cases where the existing code in the final
"else" branch would trigger a NullPointerException anyway if the value
of the switch is "null".

Commit based on refactoring by Eclipse with some manual adjustments.