Refactoring: use pattern-matching switch for cases without null
This is the biggest new feature of Java 21
and allows us to switch over types as done in this commit.
No pattern matching with deconstruction of records is done yet.

This commit changes only cases the input is non-null anyway
according to my understanding (e.g., because the code above would
already throw on null).

Commit based on refactoring by Eclipse with some manual adjustments.