Refactoring: turn BaseAlignofVisitor into switch
This needs only a little bit over half of the number of lines
and allows us to get rid of boxing.
Readability should also be better without the boilerplate of method
definitions.

The compiler still checks that all cases are handled
(note that there is no default case on the switch).

Part of #1348.