Refactoring: remove duplicate logic for setting qualifiers on CTypes
Marian pointed out that we can have one method for setting both
qualifiers and use it even when we want to change only one of them,
by passing the old value for the other (unchanged) qualifier.

This also allows us to use it for the two cases where we want to set
both qualifiers, so we do not need two copies there.