Refactoring: move all qualifier-related methods from CTypes to CType
This makes them more discoverable and easier to use.
The disadvantages are that it clutters the interface a little bit
and that the return type has to be CType now instead of the actual type
(at least if we do not want to make CType generic or add every method to
every type implementation), but both should be acceptable.
There is only a single cast that we need to add now due to the weaker
return type.