Fix NPE in LiveVariablesTransferRelation from a3bc6b06
Turns out that here is indeed one case where a null in a switch triggers
an exception that is then caught and ignored later,
so by turning the exception in a NullPointerException we produced a
crash.

The actual fix would be to properly skip initializer handling if there
is none as it was done before cc763628.
But this can be done separately, for now I am just reintroducing the
pre-a3bc6b06 to make it easier to bisect changes in the future
and such that we can see the effects from the two different fixes
individually.