Simplify CachingCanonizingCTypeVisitor in predicate analysis
It has an optional feature for removing qualifiers,
but there was no code that did not make use of it.
By making qualifier removal non-optional we can simplify the code
considerably.
The visitor now automatically removes *all* qualifiers even if more are
added in the future, not just const and volatile.
This should be the desired behavior of this visitor for example for
_Atomic types.