Support for parsing files with _Atomic qualifier with CDT
(at least as far as possible)

CDT has no support for _Atomic yet
(https://github.com/eclipse-cdt/cdt/issues/946).
So we need to apply a workaround:
We use the integrated preprocessor replace all occurrences of _Atomic
with the attribute "__CPAchecker_Atomic".
This attribute is then parsed by CDT and added to the AST nodes,
where we look for it when converting declarations.

This seems to work nicely for simple cases like
"_Atomic int a;" or "typedef _Atomic int atomic_int;".
Unfortunately, CDT also seems to have problems with attributes,
there are cases where AST nodes are supposed to have attributes
but none appear in the AST. Then our workaround does not help.
This includes for example "int * _Atomic p;",
i.e., atomic pointers.

Part of #1253.