Attempt to check for _Atomic in AST nodes where we do not handle them
There are some AST nodes where we would expect the workaround attribute
for _Atomic to occur, but it doesn't, or whether we never expect it.
This commit adds sanity checks for such cases that produce a parsing
error.
For example, if CDT adds attributes for declarators we could notice this
through these checks and could then add proper handling.