Simplify CI job auto-update-dependencies
GitLab introduced the feature that CI jobs can (if configured)
push to the git repository with the CI job token.
This means we do not need to provide and use a SSH key with push
permissions anymore.

If we add xmlstarlet to our container image we can skip installing
packages completely.