Fix return type of CFAUtils#leavingEdges(FunctionExitNode) to weaker CFANode
FunctionExitNodes can contain outgoing blank edges if distributed summary synthesis is used. So we have to weaken
the return type to avoid type errors.
This requires to adjust uses of this method that rely on the type (only FlowDepAnalysis).

Add logging to FlowDepAnalysis to hint developers to this in case it ever requires debugging