Remove ParseResult.getCFAEdges()
Using this method would be problematic in many cases, because sets of
CFAEdges are often error-prone due to equals() of CFAEdge
and because the set would become outdated as soon as the CFA is
modified.
There is only one caller of this method, where this is actually
problematic (#1397), so we inline the method there and remove it in
order to prevent further uses.