In Report.html, add shadow to popovers and graph tooltips
For better readbility. The shadow increases the contrast of tooltips on the CFA and ARG, and
of the popovers on help texts and the value assignments in the counterexample report.
Especially for the value assignments in the counterexample report, this pushes the popover
from the surrounding text and makes it easier to read.