Fix missing default resource limits in most important configs
For many years CPAchecker wanted to have and had default resource limits
in its main configs.
Unfortunately this was removed for the default config in aecf50f2.
For SV-COMP, we had a required time limit so users were forced to
specify the time limit on the command line.
But when we removed this requirement in a6de0200,
we forgot to add the default time limit.

This commit adds the default resource limits back for the default and
the current SV-COMP config.