Fully resolve position of sequence of designated initializers
Because of a misunderstanding of the semantics the ASTConverter.PositionInComposite
only resolved up to the first array subscript expression (excluding).
But designators like 'a.b[1].d' must be fully evaluated to determine
the correct position of following positional initializers.

This commit also adds a test program that demonstrates the need for this.