Fix use of offset information in preprocessor mode "auto"
If we run a preprocessor, we can only rely on line numbers,
not on offsets and column numbers.
This makes our witnesses worse
and causes non-terminating counterexample checks (#1490),
leading to a significant regression (#1466).
As a mitigation, !409 introduced a mode for running the preprocessor
only when really necessary, such that the problem occurs more rarely.

Unfortunately, the code from !409 still tells the rest of the frontend
to treat offsets and column numbers as non-reliable and ignore them,
even in the mode where the preprocessor is run only on demand.
So all the resulting problems still apply.

This commit changes it such that a CParserWithLocationMapper
that will ignore offsets and column numbers is only used
if the preprocessing actually happened (or if explicitly requested).

This finally fixes #1466.